agents/pyproject.toml
[project]
name = "neo4j-query-agent"
version = "0.1.0"
description = "Natural language query agent for Neo4j London Transport graph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.8",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.0.3",
    "neo4j>=5.25.0,<6.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
query-neo4j = "query_neo4j:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["query_neo4j.py"]

[dependency-groups]
dev = []
